requirements.txt
numpy==1.18.5
torch==1.7.1
numba
tensorboardX
easydict
pyyaml
scikit-image
tqdm
kornia==0.5.2
torchvision
wandb==0.10.13
waymo-open-dataset-tf-2-2-0